Create QiuckSort vs RadixSort
My teacher for that class insisted QuickSort was awesome. I told him radix sorting could give it a run for it's money. He had never heard of it. So I showed him the pros & cons...

s^24 30-bit entries to sort.

QuickSort: Executed in 1,809ms and used 1,404 bytes for temporary storage.

RadixSort: Executed in 189ms and used 67,110,952 bytes for temporary storage.