Create memcopy.h (ASM SSE copy routines))
Assembly holds a special place in my list of languages.

In 1993 (at the age of 13) I had the pleasure of learning from virus writers and code hackers. By the time I got to college, I was getting on the nerves of some teachers with my insistence on writing speed-critical routines in ASM.   ...And, sometimes, stuff that wasn't speed critical; just for the practice.